# apps/server-api/docs/swagger.yaml
basePath: /
definitions:
  assignment.AssignmentHintSummary:
    properties:
      createdAt:
        type: string
      hintText:
        type: string
      providerName:
        type: string
    type: object
  assignment.AssignmentProgressSummary:
    properties:
      currentRoleName:
        type: string
      currentTarget:
        type: string
      reportedAt:
        type: string
      stepSummary:
        type: string
    type: object
...
  assignment.StudentAssignmentDetail:
    properties:
      analysisStatus:
        type: string
      blockCounts:
        additionalProperties:
          type: integer
        type: object
      broadcastMessages:
        items:
          type: string
        type: array
      categoryCounts:
        additionalProperties:
          type: integer
        type: object
      description:
        type: string
      extensions:
        items:
          type: string
        type: array
      goal:
        type: string
      id:
        type: integer
      latestHint:
        $ref: '#/definitions/assignment.AssignmentHintSummary'
      latestProgress:
        $ref: '#/definitions/assignment.AssignmentProgressSummary'
      listNames:
        items:
          type: string
        type: array
      roleNames:
        items:
          type: string
        type: array
      scriptCounts:
        additionalProperties:
          type: integer
        type: object
      status:
        type: string
      teachingPoints:
        items:
          type: string
        type: array
      title:
        type: string
      variableNames:
        items:
          type: string
        type: array
    type: object
  assignment.StudentAssignmentItem:
    properties:
      analysisStatus:
        type: string
      blockCounts:
        additionalProperties:
          type: integer
        type: object
      broadcastMessages:
        items:
          type: string
        type: array
      categoryCounts:
        additionalProperties:
          type: integer
        type: object
      description:
        type: string
      extensions:
        items:
          type: string
        type: array
      goal:
        type: string
      id:
        type: integer
      listNames:
        items:
          type: string
        type: array
      roleNames:
        items:
          type: string
        type: array
      scriptCounts:
        additionalProperties:
          type: integer
        type: object
      status:
        type: string
      teachingPoints:
        items:
          type: string
        type: array
      title:
        type: string
      variableNames:
        items:
          type: string
        type: array
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  http.HintCreatedResponse:
    properties:
      assignmentId:
        example: 1
        type: integer
      hintText:
        example: 先把 Cat 的事件和移动积木串起来。
        type: string
      id:
        example: 1
        type: integer
      providerName:
        example: deepseek
        type: string
    type: object
  http.ProgressCreatedResponse:
    properties:
      assignmentId:
        example: 1
        type: integer
      currentTarget:
        example: 让角色左右移动
        type: string
      id:
        example: 1
        type: integer
      reportedAt:
        example: "2026-05-26T12:00:00Z"
        type: string
      stepSummary:
        example: 已经接上方向键事件
        type: string
    type: object
  http.ProgressReportRequestDoc:
    properties:
      currentTarget:
        example: 让 Cat 角色移动起来
        type: string
      localProjectHash:
        example: hash-1
        type: string
      reportedAt:
        example: "2026-05-26T12:00:00Z"
        type: string
      snapshot:
        $ref: '#/definitions/http.ProgressSnapshotDoc'
      stepSummary:
        example: 已经把事件积木接上了
        type: string
    required:
    - currentTarget
    - stepSummary
    type: object
  http.ProgressRoleSnapshotDoc:
    properties:
      blocks:
        items:
          type: string
        type: array
      roleName:
        example: Cat
        type: string
      roleType:
        example: sprite
        type: string
    type: object
  http.ProgressSnapshotDoc:
    properties:
      currentRoleName:
        example: Cat
        type: string
      roles:
        items:
          $ref: '#/definitions/http.ProgressRoleSnapshotDoc'
        type: array
    type: object
  http.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  http.StudentAssignmentsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/assignment.StudentAssignmentItem'
        type: array
    type: object
  http.StudentLoginRequestDoc:
    properties:
      clientType:
        example: desktop
        type: string
      password:
        example: abc12345
        type: string
      username:
        example: student01
        type: string
    required:
    - clientType
    - password
    - username
    type: object
  http.StudentMeResponse:
    properties:
      studentId:
        example: 1
        type: integer
      studentName:
        example: 小明
        type: string
      username:
        example: student01
        type: string
    type: object
...
  http.TeacherStudentsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/student.StudentItem'
        type: array
    type: object
...
  http.resetPasswordRequest:
    properties:
      newPassword:
        type: string
    required:
    - newPassword
    type: object
  student.BatchCreateInput:
    properties:
      displayName:
        type: string
      initialPassword:
        type: string
      username:
        type: string
    required:
    - displayName
    - initialPassword
    - username
    type: object
  student.BatchCreateRequest:
    properties:
      students:
        items:
          $ref: '#/definitions/student.BatchCreateInput'
        minItems: 1
        type: array
    required:
    - students
    type: object
  student.BatchCreateResult:
    properties:
      conflicts:
        items:
          type: string
        type: array
      created:
        items:
          $ref: '#/definitions/student.StudentItem'
        type: array
    type: object
  student.StudentItem:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      id:
        type: integer
      status:
        type: string
      username:
        type: string
    type: object
  student.StudentSession:
    properties:
      studentName:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
  description: OpenAPI contract for the Scratch AI teaching server.
  title: Scratch AI Server API
  version: "1.0"
paths:
...
  /api/student/assignments:
    get:
      description: List the assignments currently published and assigned to the authenticated
        student.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StudentAssignmentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List student assignments
      tags:
      - student-assignments
  /api/student/assignments/{id}:
    get:
      description: Read the assignment detail plus latest progress and hint summaries
        for the authenticated student.
      parameters:
      - description: Assignment ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/assignment.StudentAssignmentDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get student assignment detail
      tags:
      - student-assignments
  /api/student/assignments/{id}/hints:
    post:
      description: Generate the next hint from assignment analysis plus the student's
        latest reported progress.
      parameters:
      - description: Assignment ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.HintCreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request the next student hint
      tags:
      - student-assignments
  /api/student/assignments/{id}/progress:
    post:
      consumes:
      - application/json
      description: Upload the latest structured Scratch progress snapshot for one
        assigned assignment.
      parameters:
      - description: Assignment ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Progress payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.ProgressReportRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.ProgressCreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Report student progress
      tags:
      - student-assignments
  /api/student/login:
    post:
      consumes:
      - application/json
      description: Authenticate a student from the desktop client and return a bearer
        token.
      parameters:
      - description: Student login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.StudentLoginRequestDoc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/student.StudentSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Login as student
      tags:
      - student-auth
  /api/student/logout:
    post:
      description: Invalidate the current student bearer token immediately.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout student
      tags:
      - student-auth
  /api/student/me:
    get:
      description: Read the current authenticated student session details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StudentMeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current student profile
      tags:
      - student-auth
...
  /api/teacher/students:
    get:
      description: List students for the current teacher.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TeacherStudentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List students
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Create a single student for the current teacher, returning the
        same shape as batch creation.
      parameters:
      - description: Single student create payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/student.BatchCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/student.BatchCreateResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create one student
      tags:
      - students
  /api/teacher/students/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: Replace the initial or current password for one student owned by
        the current teacher.
      parameters:
      - description: Student ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: New password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/student.StudentItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reset a student password
      tags:
      - students
  /api/teacher/students/batch:
    post:
      consumes:
      - application/json
      description: Create multiple students for the current teacher and report conflicts
        without failing the entire request.
      parameters:
      - description: Batch create payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/student.BatchCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/student.BatchCreateResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch create students
      tags:
      - students
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"